# internal/model/invite.go
/*
 * Otto user service
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package model

type Invite struct {
	Code    string `json:"code"`
	Claimed bool   `json:"claimed"`
}