# internal/model/new_user.go
/*
 * Otto user service
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package model

import (
	"encoding/json"
	"net/http"
)

type NewUser struct {
	Name string `json:"name,omitempty"`

	Username string `json:"username"`

	Email string `json:"email"`

	Password string `json:"password"`

	InviteCode string `json:"invite_code"`
}

func DecodeRequestToNewUser(r *http.Request) *NewUser {
	decoder := json.NewDecoder(r.Body)
	var data *NewUser
	err := decoder.Decode(&data)
	if err != nil {
		panic(err)
	}
	return data
}