# internal/model/session_refresh.go
/*
 * Otto user service
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package model

import (
	"encoding/json"
	"net/http"
)

// SessionRefresh struct for SessionRefresh
type SessionRefresh struct {
	Token string `json:"token"`
}

func DecodeRequestToSessionRefresh(r *http.Request) *SessionRefresh {
	decoder := json.NewDecoder(r.Body)
	var data *SessionRefresh
	err := decoder.Decode(&data)
	if err != nil {
		panic(err)
	}
	return data
}